Twitter users are now processed using new user and video services - in Casandra videos and users are represented as integers.

+++ src/main/scala/ztis/user_video_service/UserServiceActor.scala
@@ -10,7 +10,7 @@ import ztis.user_video_service.persistence.{Metadata, MetadataRepository, UnitOf
 
 object UserServiceActor {
 
-  case class RegisterTwitterUser(externalUserName: String, externalUserID: Int)
+  case class RegisterTwitterUser(externalUserName: String, externalUserID: Long)
 
   case class RegisterWykopUser(externalUserName: String)
